requirements.txt
dropbox~=12.0.1
python-dotenv~=1.0.1
requests~=2.32.3
beautifulsoup4~=4.12.2
pythumb~=1.1.0
